build: add safety net for mac_deployment_target
As of #28480 we now dynamically determine the LSMinimumSystemVersion
value as part of the build process.  To avoid this changing and no one
realizing we now have this assert which will trip during a Chromium
upgrade if they bump the minimum supported macOS version so we can
update our documentation appropriately.